Seperating chunk utility into naming and copying, adding in JSON functionality for #20
