+++ ChunkTypes.vcxproj
@@ -177,13 +177,14 @@
   <ItemGroup>
     <ClInclude Include="include/BaseChunk.h" />
     <ClInclude Include="include/ChunkTypes.h" />
-    <ClInclude Include="include/ChunkTypesUtility.h" />
+    <ClInclude Include="include\ChunkTypesNamingUtility.h" />
     <ClInclude Include="include/TimeChunk.h" />
     <ClInclude Include="include/SessionModeTypes.h" />
     <ClInclude Include="include/UDPChunk.h" />
     <ClInclude Include="include/WAVChunk.h" />
     <ClInclude Include="include/JSONChunk.h" />
     <ClInclude Include="include/WatchdogChunk.h" />
+    <ClInclude Include="include\ChunkDuplicator.h" />
     <ClInclude Include="include\ChunkToJSONConverter.h" />
   </ItemGroup>
   <ItemGroup>
