+++ ChunkTypes.vcxproj.filters
@@ -21,7 +21,7 @@
     <ClInclude Include="include/ChunkTypes.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="include/ChunkTypesUtility.h">
+    <ClInclude Include="include\ChunkTypesNamingUtility.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="include/TimeChunk.h">
@@ -43,6 +43,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="include\ChunkToJSONConverter.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="include\ChunkDuplicator.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
