+++ include/ChunkTypesNamingUtility.h
@@ -16,12 +16,12 @@
 /**
  * @brief Class that facillitates conversion between emunerated chunk type and string
  */
-class ChunkTypesUtility
+class ChunkTypesNamingUtility
 {
 public:
 
-    ChunkTypesUtility() {};
-    ~ChunkTypesUtility() {};
+    ChunkTypesNamingUtility() {};
+    ~ChunkTypesNamingUtility() {};
 
     static std::string toString(ChunkType eChunkType)
     {
@@ -56,30 +56,12 @@ public:
         switch (u32ChunkType)
         {
         case 1: return ChunkType::ChunkBase;
-        case 2:  return ChunkType::UDPChunk;
+        case 2: return ChunkType::UDPChunk;
         case 3: return ChunkType::TimeChunk;
-        case 4:  return ChunkType::WAVChunk;
-        case 5:  return ChunkType::WatchdogChunk;
+        case 4: return ChunkType::WAVChunk;
+        case 5: return ChunkType::WatchdogChunk;
 
         default: return ChunkType::ChunkBase; // error
-        }
-    }
-
-    static std::shared_ptr<BaseChunk> DuplicateDerivedChunk(std::shared_ptr<BaseChunk> pBaseChunk)
-    {
-        switch (pBaseChunk->GetChunkType())
-        {
-            case ChunkType::ChunkBase: return std::make_shared<BaseChunk>(pBaseChunk);
-            case ChunkType::UDPChunk:  return std::make_shared<UDPChunk>(std::static_pointer_cast<UDPChunk>(pBaseChunk));
-            case ChunkType::TimeChunk: return std::make_shared<TimeChunk>(std::static_pointer_cast<TimeChunk>(pBaseChunk));
-            case ChunkType::WAVChunk:  return std::make_shared<WAVChunk>(std::static_pointer_cast<WAVChunk>(pBaseChunk));
-            case ChunkType::WatchdogChunk:  return std::make_shared<WatchdogChunk>(std::static_pointer_cast<WatchdogChunk>(pBaseChunk));
-
-            default:
-            {
-                std::cout << std::string(__FUNCTION__) + ": Unknown chunk type \n";
-                return std::make_shared<BaseChunk>(BaseChunk(pBaseChunk));
-            }
         }
     }
 };
